+++ src/public/load-config.js
@@ -1,0 +1,2 @@
+import { loadConfig } from '../lib/load-config'
+export default loadConfig