+++ src/util/resolveConfigPath.js
@@ -1,6 +1,13 @@
 import fs from 'fs'
 import path from 'path'
 
+const defaultConfigFiles = [
+  './tailwind.config.js',
+  './tailwind.config.cjs',
+  './tailwind.config.mjs',
+  './tailwind.config.ts',
+]
+
 function isObject(value) {
   return typeof value === 'object' && value !== null
 }
@@ -43,7 +50,11 @@ export default function resolveConfigPath(pathOrConfig) {
   }
 
   // require('tailwindcss')
-  for (const configFile of ['./tailwind.config.js', './tailwind.config.cjs']) {
+  return resolveDefaultConfigPath()
+}
+
+export function resolveDefaultConfigPath() {
+  for (const configFile of defaultConfigFiles) {
     try {
       const configPath = path.resolve(configFile)
       fs.accessSync(configPath)