+++ stubs/config.simple.js
@@ -1,0 +1,7 @@
+module.exports = {
+  content: [],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}