+++ stubs/tailwind.config.ts
@@ -1,0 +1,3 @@
+import type { Config } from 'tailwindcss'
+
+export default __CONFIG__ satisfies Config